+++ src/main/java/racingcar/view/ConsoleView.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.view;
 
 
 import java.util.Scanner;
