+++ src/test/java/racingcar/RandomGeneratorTest.java
@@ -1,10 +1,12 @@
 package racingcar;
 
+import org.junit.jupiter.api.DisplayNameGeneration;
+import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-
+@DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
 class RandomGeneratorTest {
 
     @Test
